Move all auth stuff to authHelpers
