+++ src/containers/App/LogoutFederated.js
@@ -1,5 +1,4 @@
-/*
- * Part of NDLA editorial-frontend.
+/**
  * Copyright (C) 2017 -present, NDLA
  *
  * This source code is licensed under the GPLv3 license found in the
