+++ src/reducers/authenticated.js
@@ -14,5 +14,4 @@ export default handleActions({
     next: (state, action) => action.payload,
     throw: state => state,
   },
-  // LOGOUT: () => false,
 }, false);
