--- src/util/ifAuthenticated.js
@@ -1,8 +1,0 @@
-export default function ifAuthenticated(authenticated, cb) {
-  return (...args) => {
-    if (authenticated) {
-      return cb ? cb(...args) : undefined;
-    }
-    return undefined;
-  };
-}
