+++ src/util/jwtHelper.js
@@ -8,26 +8,9 @@
 
  import decode from 'jwt-decode';
 
- export function getTokenExpiration(token) {
-   const decoded = decode(token);
-   return decoded.exp;
- }
- export function getTokenIssuedAt(token) {
-   const decoded = decode(token);
-   return decoded.iat;
- }
-
  export function expiresIn(token) {
    const decoded = decode(token);
-   return decoded.exp - decoded.iat - 60;
+   return decoded.exp - decoded.iat - 60; // Add 60 second buffer
  }
 
- export function isTokenExpired(token) {
-   return getTokenExpiration(token) - getTokenIssuedAt(token) < 0;
- }
-
- export const decodeIdToken = idToken => decode(idToken);
-
- export function getTimeToUpdateInMs(token) {
-   return (getTokenExpiration(token) - getTokenIssuedAt(token) - (60 * 5)) * 1000; // Removes 5 minutes from time to update
- }
+ export const decodeToken = idToken => decode(idToken);
